don't clean things that don't exist and no longer create spurious build statuses

+++ misc/o2/o2-unbuild.sh
@@ -7,4 +7,4 @@ if cd 01bld; then
     rm -fr *-x-* lists/*
     cd ..
 fi
-(cd 01tmp && rm -fr *.*)
+[ -d 01tmp ] && (cd 01tmp && rm -fr *.*)
